`zedrouter` move metadata handler to separate file
This is first part of refactoring metadata server
to support configuration services.
